requirements.txt
requests>=2.31.0
click>=8.1.7
websockets>=12.0
python-dotenv>=1.0.0